+++ tests/src/TrySendDiscordError.php
@@ -1,11 +1,17 @@
 <?php
 
+use Illuminate\Support\Facades\Http;
+use TomatoPHP\LaravelDiscordErrorTracker\Services\DiscordServices;
+
 use function PHPUnit\Framework\assertTrue;
 
 it('can send discord error', function () {
-    $exception = new \Exception('Test Exception');
-    $discordServices = new \TomatoPHP\LaravelDiscordErrorTracker\Services\DiscordServices;
+    Http::fake();
+
+    $exception = new Exception('Test Exception');
+    $discordServices = new DiscordServices;
     $response = $discordServices->handler($exception);
 
     assertTrue($response);
+    Http::assertSentCount(1);
 });